requirements.txt
dateutils
lxml
newspaper
python-jsonrpc
pymongo
pika
requests
redis
sklearn
